Restructure unique op test generation
