+++ forge/test/conftest.py
@@ -101,9 +101,6 @@ def clear_forge():
 
 def pytest_addoption(parser):
     parser.addoption(
-        "--generate-op-tests", action="store_true", default=False, help="Generate op tests for the given model"
-    )
-    parser.addoption(
         "--generate-unique-op-tests",
         action="store_true",
         default=False,
@@ -185,11 +182,9 @@ def initialize_global_compiler_configuration_based_on_pytest_args(pytestconfig):
     """
     compiler_cfg = _get_global_compiler_config()
 
-    compiler_cfg.tvm_generate_op_tests = pytestconfig.getoption("--generate-op-tests")
-
     compiler_cfg.tvm_generate_unique_op_tests = pytestconfig.getoption("--generate-unique-op-tests")
 
-    if compiler_cfg.tvm_generate_op_tests or compiler_cfg.tvm_generate_unique_op_tests:
+    if compiler_cfg.tvm_generate_unique_op_tests:
         # For running standalone tests, we need to retain the generated python files
         # together with stored model parameters
         compiler_cfg.retain_tvm_python_files = True
